Refactor to Strategy Pattern
Changes the code to strategy pattern. Supports Kyber (hybrid or pq-only) using P-384 and Kyber768 parameters. I can justify strategy due to the possibility of switching from hybrid to pq-only without great impacts in the code; and also due to the possibility of creating more hybrid modes (e.g., Hybrid Confidentiality mentioned in the KEMTLS-PDK paper; or using ECIES instead of ECDH; but this is not supported here). However, one problem is repeating code between strategies. A second problem is the java KAS API requiring UKM (User keying material). In this example, UKM is defined outside concrete strategies so maybe the best approach would be creating additional parameters or check if a zero-ed byte vector as ukm is secure for ephemeral establishment.